+++ schema/schema.go
@@ -81,12 +81,11 @@ type BuiltinAuthProvider struct {
 
 // ExperimentalFeatures description: Experimental features to enable or disable. Features that are now enabled by default are marked as deprecated.
 type ExperimentalFeatures struct {
-	CanonicalURLRedirect   string `json:"canonicalURLRedirect,omitempty"`
-	ConfigVars             string `json:"configVars,omitempty"`
-	HostSurveysLocally     string `json:"hostSurveysLocally,omitempty"`
-	JumpToDefOSSIndex      string `json:"jumpToDefOSSIndex,omitempty"`
-	MultipleAuthProviders  string `json:"multipleAuthProviders,omitempty"`
-	SearchTimeoutParameter string `json:"searchTimeoutParameter,omitempty"`
+	CanonicalURLRedirect  string `json:"canonicalURLRedirect,omitempty"`
+	ConfigVars            string `json:"configVars,omitempty"`
+	HostSurveysLocally    string `json:"hostSurveysLocally,omitempty"`
+	JumpToDefOSSIndex     string `json:"jumpToDefOSSIndex,omitempty"`
+	MultipleAuthProviders string `json:"multipleAuthProviders,omitempty"`
 }
 type GitHubConnection struct {
 	Certificate                 string   `json:"certificate,omitempty"`